* main:
  Refactor tests
  Add `CONTRIBUTING.md`
  Rename `NewFromUrl` to `NewFromURL`
  Add support for large file upload

+++ api/admin/metadata_fields_test.go
@@ -1,17 +1,19 @@
-package admin
+package admin_test
 
 import (
 	"log"
 	"testing"
 	"time"
 
+	"github.com/cloudinary/cloudinary-go/api/admin"
 	"github.com/cloudinary/cloudinary-go/api/admin/metadata"
+	"github.com/cloudinary/cloudinary-go/internal/cldtest"
 )
 
 var metadataField = metadata.Field{
 	Type:         metadata.SetFieldType,
-	ExternalID:   "go_color_id_" + testSuffix,
-	Label:        "GoColors" + testSuffix,
+	ExternalID:   cldtest.UniqueID("go_color_id_"),
+	Label:        cldtest.UniqueID("GoColors"),
 	Mandatory:    true,
 	DefaultValue: []string{"go_color1", "go_color2"},
 	DataSource:   dataSource1,
@@ -64,9 +66,9 @@ func TestAdmin_AddMetadataField(t *testing.T) {
 }
 
 func TestAdmin_UpdateMetadataField(t *testing.T) {
-	metadataField.Label = "GoUpdatedColors" + testSuffix
+	metadataField.Label = cldtest.UniqueID("GoUpdatedColors")
 
-	resp, err := adminApi.UpdateMetadataField(ctx, UpdateMetadataFieldParams{
+	resp, err := adminApi.UpdateMetadataField(ctx, admin.UpdateMetadataFieldParams{
 		FieldExternalId: metadataField.ExternalID,
 		Field:           metadataField,
 	})
@@ -89,7 +91,7 @@ func TestAdmin_ListMetadataFields(t *testing.T) {
 }
 
 func TestAdmin_MetadataFieldByFieldId(t *testing.T) {
-	params := MetadataFieldByFieldIdParams{FieldExternalId: metadataField.ExternalID}
+	params := admin.MetadataFieldByFieldIdParams{FieldExternalId: metadataField.ExternalID}
 	resp, err := adminApi.MetadataFieldByFieldId(ctx, params)
 
 	if err != nil || resp.ExternalID != metadataField.ExternalID {
@@ -98,7 +100,7 @@ func TestAdmin_MetadataFieldByFieldId(t *testing.T) {
 }
 
 func TestAdmin_UpdateMetadataFieldDataSource(t *testing.T) {
-	resp, err := adminApi.UpdateMetadataFieldDataSource(ctx, UpdateMetadataFieldDataSourceParams{
+	resp, err := adminApi.UpdateMetadataFieldDataSource(ctx, admin.UpdateMetadataFieldDataSourceParams{
 		FieldExternalId: metadataField.ExternalID,
 		DataSource:      dataSource2,
 	})
@@ -109,7 +111,7 @@ func TestAdmin_UpdateMetadataFieldDataSource(t *testing.T) {
 }
 
 func TestAdmin_DeleteDataSourceEntries(t *testing.T) {
-	resp, err := adminApi.DeleteDataSourceEntries(ctx, DeleteDataSourceEntriesParams{
+	resp, err := adminApi.DeleteDataSourceEntries(ctx, admin.DeleteDataSourceEntriesParams{
 		FieldExternalId:    metadataField.ExternalID,
 		EntriesExternalIDs: []string{"go_color3", "go_color4"},
 	})
@@ -120,7 +122,7 @@ func TestAdmin_DeleteDataSourceEntries(t *testing.T) {
 }
 
 func TestAdmin_RestoreMetadataFieldDataSource(t *testing.T) {
-	resp, err := adminApi.RestoreDatasourceEntries(ctx, RestoreDatasourceEntriesParams{
+	resp, err := adminApi.RestoreDatasourceEntries(ctx, admin.RestoreDatasourceEntriesParams{
 		FieldExternalId:    metadataField.ExternalID,
 		EntriesExternalIDs: []string{"go_color3", "go_color4"},
 	})
@@ -131,7 +133,7 @@ func TestAdmin_RestoreMetadataFieldDataSource(t *testing.T) {
 }
 
 func TestAdmin_DeleteMetadataField(t *testing.T) {
-	resp, err := adminApi.DeleteMetadataField(ctx, DeleteMetadataFieldParams{FieldExternalId: metadataField.ExternalID})
+	resp, err := adminApi.DeleteMetadataField(ctx, admin.DeleteMetadataFieldParams{FieldExternalId: metadataField.ExternalID})
 
 	if err != nil || resp.Message != "ok" {
 		t.Error(err, resp)
@@ -139,17 +141,17 @@ func TestAdmin_DeleteMetadataField(t *testing.T) {
 }
 
 var mdIDs = map[string]string{
-	"enum": "go_distinct_color_id_" + testSuffix,
-	"int":  "go_17_integer_id_" + testSuffix,
-	"str":  "go_string_id_" + testSuffix,
-	"date": "go_date_id_" + testSuffix,
+	"enum": cldtest.UniqueID("go_distinct_color_id_"),
+	"int":  cldtest.UniqueID("go_17_integer_id_"),
+	"str":  cldtest.UniqueID("go_string_id_"),
+	"date": cldtest.UniqueID("go_date_id_"),
 }
 
 func TestAdmin_AddMetadataFields(t *testing.T) {
 	var integerMetadataField = metadata.Field{
 		Type:       metadata.IntegerFieldType,
 		ExternalID: mdIDs["int"],
-		Label:      "Go17Integer" + testSuffix,
+		Label:      cldtest.UniqueID("Go17Integer"),
 		Validation: metadata.AndValidation(
 			[]interface{}{
 				metadata.GreaterThanValidation(17, true),
@@ -160,7 +162,7 @@ func TestAdmin_AddMetadataFields(t *testing.T) {
 	var stringMetadataField = metadata.Field{
 		Type:         metadata.StringFieldType,
 		ExternalID:   mdIDs["str"],
-		Label:        "GoString" + testSuffix,
+		Label:        cldtest.UniqueID("GoString"),
 		DefaultValue: "Gopher",
 		Validation:   metadata.StringLengthValidation(2, 6),
 	}
@@ -168,7 +170,7 @@ func TestAdmin_AddMetadataFields(t *testing.T) {
 	var dateMetadataField = metadata.Field{
 		Type:         metadata.DateFieldType,
 		ExternalID:   mdIDs["date"],
-		Label:        "GoDate" + testSuffix,
+		Label:        cldtest.UniqueID("GoDate"),
 		DefaultValue: time.Now().Format("2006-01-02"),
 		Validation:   metadata.GreaterThanValidation(time.Now().AddDate(0, 0, -1), false),
 	}
@@ -176,7 +178,7 @@ func TestAdmin_AddMetadataFields(t *testing.T) {
 	var enumMetadataField = metadata.Field{
 		Type:       metadata.EnumFieldType,
 		ExternalID: mdIDs["enum"],
-		Label:      "GoDistinctColors" + testSuffix,
+		Label:      cldtest.UniqueID("GoDistinctColors"),
 		DataSource: dataSource1,
 	}
 
@@ -205,7 +207,7 @@ func TestAdmin_AddMetadataFields(t *testing.T) {
 //FIXME; find a good library with a proper TearDown method
 func TestAdmin_MetadataFieldsCleanup(t *testing.T) {
 	for _, extID := range mdIDs {
-		resp, err := adminApi.DeleteMetadataField(ctx, DeleteMetadataFieldParams{FieldExternalId: extID})
+		resp, err := adminApi.DeleteMetadataField(ctx, admin.DeleteMetadataFieldParams{FieldExternalId: extID})
 		if err != nil || resp.Message != "ok" {
 			log.Println(err, resp)
 		}
